Restructure components directory
