+++ src/components/Root.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { HashRouter as Router, Route } from 'react-router-dom';
 
-import HomePage from './HomePage/HomePage';
-import TestPage from './TestPage/TestPage';
+import HomePage from './pages/home/HomePage';
+import TestPage from './pages/test/TestPage';
 import ApplicationBar from './ApplicationBar';
 
 export default class Root extends React.Component {
